Allow to set different globals per locale
This changes the globals that when it has have a `default` key it can
have locale aware globals. For example:

```
I18n.config.globals = {
  salutation: 'Bye bye',
  hello: 'Hello',
  de: {
    salutation: 'Tschüss'
  },
  fr: {
    salutation: 'Salut'
  }
}
```

Depending on the locale, the global variable `salutation` will be
different. The default value is used as a fallback, so that `hello` will
be available in each language, too.